feat(site): route app home to studio
