+++ site/style.css
@@ -18,6 +18,18 @@
 
 html, body { margin: 0; padding: 0; }
 
+.sr-only {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  margin: -1px;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  white-space: nowrap;
+  border: 0;
+}
+
 body {
   background: var(--vellum);
   color: var(--ink);
@@ -174,6 +186,56 @@ body {
   line-height: 1.45;
 }
 
+/* PROMPT HANDOFF */
+.prompt-handoff {
+  display: grid;
+  grid-template-columns: minmax(220px, 0.72fr) minmax(320px, 1fr);
+  gap: 28px;
+  align-items: end;
+  margin: 0 0 88px;
+}
+.section-kicker {
+  display: block;
+  font-family: 'JetBrains Mono', ui-monospace, monospace;
+  font-size: 11px;
+  color: var(--blueprint);
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+  margin-bottom: 10px;
+}
+.prompt-handoff-headline {
+  font-family: 'Source Serif 4', Georgia, serif;
+  font-size: clamp(32px, 4.5vw, 54px);
+  font-weight: 500;
+  line-height: 1;
+  letter-spacing: 0;
+  margin: 0;
+}
+.prompt-handoff-form {
+  display: grid;
+  gap: 12px;
+}
+.prompt-handoff-form textarea {
+  width: 100%;
+  min-height: 108px;
+  resize: vertical;
+  border: 1px solid var(--rule);
+  border-radius: 8px;
+  background: var(--paper);
+  color: var(--ink);
+  padding: 15px 16px;
+  font: inherit;
+  font-size: 15px;
+  line-height: 1.45;
+}
+.prompt-handoff-form textarea:focus {
+  outline: none;
+  border-color: var(--blueprint);
+}
+.prompt-handoff-form .btn {
+  justify-content: center;
+}
+
 /* CODE */
 .code-block { margin: 0 0 100px; max-width: 720px; margin-left: auto; margin-right: auto; }
 .code-label {
@@ -288,6 +350,7 @@ body {
   .capabilities { grid-template-columns: 1fr; margin-bottom: 60px; }
   .capability { border-right: 0; border-bottom: 1px solid var(--rule); }
   .capability:last-child { border-bottom: 0; }
+  .prompt-handoff { grid-template-columns: 1fr; margin-bottom: 60px; }
   .code { font-size: 12px; padding: 20px; }
   .footer { flex-direction: column; gap: 12px; }
   .signup { margin: 40px 0 60px; padding: 24px 18px; }
