+++ src/studio/routes/index.test.ts
@@ -1,19 +1,22 @@
 import { readFileSync } from 'node:fs';
 import { describe, expect, it } from 'vitest';
 
-/**
- * Landing-page contract: the kernelcad.com landing route MUST render the
- * three signals-collecting sections in order. Each one has been silently
- * dropped at least once during the funnel iteration. Future PRs that
- * remove any of them have to either delete this test (visible diff in
- * review) or fail CI.
- *
- *   1. PromptBox      — primary CTA, "describe a part" textarea
- *   2. GallerySection — "Built with kernelCAD" curated tiles
- *   3. EmailSignup    — fallback signal collection, bottom of page
- */
-describe('Landing page (src/studio/routes/index.tsx)', () => {
+describe('App root route (src/studio/routes/index.tsx)', () => {
   const source = readFileSync('src/studio/routes/index.tsx', 'utf8');
+
+  it('opens the Studio shell at app root instead of the generation prompt funnel', () => {
+    expect(source).toMatch(/import\s+App\s+from\s+['"]\.\.\/App['"]/);
+    expect(source).toMatch(/<App\s*\/>/);
+    expect(source).not.toMatch(/<PromptBox\b/);
+  });
+});
+
+/**
+ * Generate-page contract: the prompt funnel remains available, but it no
+ * longer owns app root. It keeps the three signal sections in order.
+ */
+describe('Generate page (src/studio/routes/generate.tsx)', () => {
+  const source = readFileSync('src/studio/routes/generate.tsx', 'utf8');
 
   it('imports the three section components', () => {
     expect(source).toMatch(/from\s+['"]\.\.\/\.\.\/funnel\/components\/PromptBox['"]/);
